Changing the UserLoginSignup and now it is the home page but still working on redirection!
