+++ Team4of5/src/Team4of5_App/BugTracking/BugTracking.js
@@ -1,4 +1,3 @@
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 
@@ -7,7 +6,7 @@ class BugTracking extends React.Component {
     constructor(props) {
         super(props);
     }
-    
+
     render() {
         return (
             <h1>BugTracking</h1>
